hidapi-async: update libusb/hidapi submodule of the hidapi dep
This fixes https://github.com/digitalbitbox/bitbox-bridge/issues/43 -
a regression where libusb/hidapi would report the interface number of
both (HWW and U2F) as 0 instead of 0 and 1 on macOS 13.3.

The fix is to pull the upstream fix:
https://github.com/libusb/hidapi/pull/534.

hidapi-rs was forked to benma/hidapi-rs to pull in the latest
libusb/hidapi that contains the fix.

This also led me down a rabbit-hole of updating the Docker base image
and all sorts of dependencies.

Code changes are clippy fixes.

libusb dep was changed to libudev, as hidapi-rs switched to the hidraw
hidapi backend by default (was libusb backend before) in
https://github.com/ruabmbua/hidapi-rs/commit/bbb7f3eb4ef91f71cbe2b39b4960019aa110112e.

Docker base image updated from 16.04 to 20.04 as as 16.04 and 18.04
reached end-of-life.